intention/requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-jquery==3.1.0
django-tz-detect==0.2.9
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
numpy==1.16.1
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
